+++ src/Foundation/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'Vanilo Foundation',
-    'version' => '5.2.0',
+    'version' => '5.2.1',
 ];