+++ Changelog.md
@@ -11,6 +11,7 @@
 - Dropped MySQL 5.7 support (it still works, but it is no longer included in the CI tests)
 - Dropped MySQL 8.0 support (it still works, but it is no longer included in the CI tests)
 - Added explicit MySQL 8.4 support (being actively tested against in the CI)
+- Added the konekt/enum package dependency (v4.2) to the contracts module
 - Changed the Postgres testing policy to use v15 and v18
 
 ## 5.x Series