+++ Changelog.md
@@ -18,6 +18,7 @@
 - Added `Taxable` and `Buyable` to the ShippingMethod model (Foundation)
 - Added support for shipping line items, where the Buyable is a ShippingMethod (opt-in via the `foundation.use_shipping_lines` config flag)
 - Added `usage_count` and `last_usage_at` fields to shipping methods
+- Added the `feature()`, `feature_is_enabled()`, and `feature_is_disabled()` helpers
 - Changed the Postgres testing policy to use v15 and v18
 - Changed the implementation of the `TaxCalculators` to comply with the Extend's Registry interface)